사용자 용량 클시 integer -> double로 수정

+++ src/main/java/egovframework/ezEKP/ezWebFolder/web/EzWebFolderGWController.java
@@ -334,7 +334,7 @@ public class EzWebFolderGWController {
 			}
 			else {
 				double totalCapByBytes = Double.parseDouble(userCapacity.getTotalCapacity()) * 10737418.24;
-				userCapacity.setUsedRate((int)(Integer.parseInt(userCapacity.getTotalUsed())/totalCapByBytes));
+				userCapacity.setUsedRate((int)(Double.parseDouble(userCapacity.getTotalUsed())/totalCapByBytes));
 			}
 			
 			result.put("data", userCapacity);
